+++ CriticalMass/NetworkLayer.swift
@@ -11,4 +11,5 @@ protocol NetworkLayer {
     func get<T: Decodable>(with url: URL, decodable: T.Type, completion: @escaping (T?) -> Void)
     func get<T: Decodable>(with url: URL, decodable: T.Type, customDateFormatter: DateFormatter?, completion: @escaping (T?) -> Void)
     func post<T: Decodable>(with url: URL, decodable: T.Type, bodyData: Data, completion: @escaping (T?) -> Void)
+    func cancelActiveRequestsIfNeeded()
 }